Problem number 2
tkinter was not able to find the attribute "show frame" due controller not being defined. In order to fix this, I took the old code for OXO and re-wrote by adding each sub routines into defined functions and connected it to the main coin flip code by editing the "start_oxo_game" function to import the code from OXO Terminal.